Refactor rename template part modal after dataview changes
